feat: presigned url 발급 api 구현 및 balance 등록 api request 수정

+++ src/main/kotlin/com/balancemania/api/client/mediapipe/model/MediaPipeImgAnalysisResponse.kt
@@ -2,20 +2,20 @@ package com.balancemania.api.client.mediapipe.model
 
 data class MediaPipeImgAnalysisResponse(
     /** 정면 어깨 각도 */
-    val frontShoulderAngle: Long,
+    val frontShoulderAngle: Float,
 
     /** 정면 골반 각도 */
-    val frontPelvisAngle: Long,
+    val frontPelvisAngle: Float,
 
     /** 정면 무릎 각도 */
-    val frontKneeAngle: Long,
+    val frontKneeAngle: Float,
 
     /** 정면 발목 각도 */
-    val frontAnkleAngle: Long,
+    val frontAnkleAngle: Float,
 
     /** 측면 목 각도 */
-    val sideNeckAngle: Long,
+    val sideNeckAngle: Float,
 
     /** 측면 신체 각도 */
-    val sideBodyAngle: Long,
+    val sideBodyAngle: Float,
 )
